+++ data-raw/dat.franchini2012.r
@@ -1,7 +1,6 @@
 # Franchini AJ, Dias S, Ades AE, Jansen JP, Welton NJ (2012).
 # Accounting for correlation in network meta-analysis with multi-arm
-# trials.
-# Research Synthesis Methods, 3, 142-60.
+# trials. Research Synthesis Methods, 3, 142-60.
 
 dat.franchini2012 <- read.csv("data-raw/dat.franchini2012.txt")
 save(dat.franchini2012, file="data/dat.franchini2012.rda")