+++ NEWS.md
@@ -1,7 +1,7 @@
-# metadat 1.1-0 (2021-10-13)
+# metadat 1.1-0 (2022-01-10)
 
 - added some more info to `dat.knapp2017`
-- added `dat.lehmann2018`, `dat.michael2013`, `dat.mccurdy2020`, `dat.baker2009`, and `dat.bakdash2021`
+- added `dat.lehmann2018`, `dat.michael2013`, `dat.mccurdy2020`, `dat.baker2009`, `dat.bakdash2021`, `dat.dong2013`, `dat.franchini2012`, `dat.gurusamy2011`, `dat.linde2015`, `dat.linde2016`, and `dat.woods2010`
 - changed concept term 'mean differences' to 'raw mean differences' (to better distinguish it from the concept 'standardized mean differences')
 
 # metadat 1.0-0 (2021-08-20)